refactor(ia-t2): collapse booking detail into one role-aware component
