+++ src/app/(protected)/guide/bookings/[bookingId]/page.tsx
@@ -1,5 +1,5 @@
 import type { Metadata } from "next";
-import { GuideBookingDetailScreen } from "@/features/guide/components/bookings/guide-booking-detail-screen";
+import { BookingDetailScreen } from "@/features/bookings/components/booking-detail-screen";
 
 export const metadata: Metadata = {
   title: "Бронирование",
@@ -11,6 +11,6 @@ export default async function GuideBookingDetailPage({
   params: Promise<{ bookingId: string }>;
 }) {
   const { bookingId } = await params;
-  return <GuideBookingDetailScreen bookingId={bookingId} />;
+  return <BookingDetailScreen viewerRole="guide" bookingId={bookingId} />;
 }
